+++ src/main/java/com/vigiaenchente/api/repository/AddressRepository.java
@@ -10,13 +10,13 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface AddressRepository extends JpaRepository<Address, Long> {
+public interface AddressRepository extends JpaRepository<Address, Integer> {
 
     /**
      * Find address by user ID
      */
     @Query("SELECT a FROM Address a WHERE a.user.id = :userId")
-    Optional<Address> findByUserId(@Param("userId") Long userId);
+    Optional<Address> findByUserId(@Param("userId") Integer userId);
 
     /**
      * Find all addresses in a specific city
@@ -32,5 +32,5 @@ public interface AddressRepository extends JpaRepository<Address, Long> {
      * Delete address by user ID
      */
     @Query("DELETE FROM Address a WHERE a.user.id = :userId")
-    void deleteByUserId(@Param("userId") Long userId);
+    void deleteByUserId(@Param("userId") Integer userId);
 }
