+++ src/main/java/com/vigiaenchente/api/repository/UserRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
     /**
      * Find user by email
@@ -35,7 +35,7 @@ public interface UserRepository extends JpaRepository<User, Long> {
      * Find user with address loaded
      */
     @Query("SELECT u FROM User u LEFT JOIN FETCH u.address WHERE u.id = :id")
-    Optional<User> findByIdWithAddress(@Param("id") Long id);
+    Optional<User> findByIdWithAddress(@Param("id") Integer id);
 
     /**
      * Find user by email with address loaded
