+++ src/main/java/com/vigiaenchente/api/service/UserService.java
@@ -15,9 +15,8 @@ public class UserService {
     private final UserRepository userRepository;
 
     @Transactional(readOnly = true)
-    public UserProfileResponse getUserProfile(Long userId) {
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new ResourceNotFoundException("Usuário não encontrado"));
+    public UserProfileResponse getUserProfile(Integer userId) {
+        User user = userRepository.findById(userId).orElseThrow(() -> new ResourceNotFoundException("Usuário não encontrado"));
 
         return UserProfileResponse.fromEntity(user);
     }
